:building_construction: Move Search to Independent Module
