+++ src/Shared/OutcomeFilter.ts
@@ -1,0 +1,7 @@
+export interface OutcomeFilter {
+    [key: string]: string;
+    text?: string;
+    source?: string;
+    name?: string;
+    date?: string;
+}
